After some consideration, use AverageVoltage for Vin and the hood pot. If we could sample Vin regularly we might want fast updates, but since we only get to use it when initializing other components we'd rather have some hysteresis.
The hood pot is a used in direct reference to Vin, so we'd like updates on the same scale. Plus the hood should be slow moving with respect to our voltage changes.

But keep the transient voltage for the accelerometer; that system is not supposed to be stable and we want genuinely new data on every cycle.